organized functions, coaches can now add players, league owners can now add teams and assign a coach to it
