+++ README.md
@@ -1,1 +1,12 @@
 # cpsc431-proj1
+## 'andre' branch
+
+### make sure to check out 'config.php'
+Define your base url (since our paths after localhost might be different)
+
+### Currently not using:
+processAddressUpdate.php
+processStatsitcUpdate.php
+home_page.php
+Address.php
+PlayerStatistic.php
